+++ tools/rendercli/rendercli.cpp
@@ -271,6 +271,12 @@ namespace {
         << unsignedValue(batching, "intersectionEstimatedAnyHitReadbackBytes")
         << " intersection_estimated_query_transfer_bytes="
         << unsignedValue(batching, "intersectionEstimatedQueryTransferBytes")
+        << " intersection_backend_upload_worker_ms="
+        << batching.value("intersectionBackendUploadWorkerSeconds").toDouble() * 1000.0
+        << " intersection_backend_kernel_worker_ms="
+        << batching.value("intersectionBackendKernelWorkerSeconds").toDouble() * 1000.0
+        << " intersection_backend_readback_worker_ms="
+        << batching.value("intersectionBackendReadbackWorkerSeconds").toDouble() * 1000.0
         << " intersection_rays=" << unsignedValue(batching, "intersectionRaysSubmitted")
         << " closest_hit_queries=" << unsignedValue(batching, "closestHitQueries")
         << " any_hit_queries=" << unsignedValue(batching, "anyHitQueries")